Fix incorrect use of scoped enumerations in format string
Scoped enums do not get automatically promoted to their underlying type,
so these uses have undefined behavior and Clang recently started warning
about it.

Bug: chromium:1456289
Reviewed-on: https://swiftshader-review.googlesource.com/c/SwiftShader/+/71788
Kokoro-Result: kokoro <noreply+kokoro@google.com>
Commit-Queue: Geoff Lang <geofflang@google.com>
Tested-by: Geoff Lang <geofflang@google.com>

+++ src/Reactor/LLVMJIT.cpp
@@ -155,7 +155,7 @@ static void *getTLSAddress(void *control)
 	case MSanTLS::origin: return reinterpret_cast<void *>(&__msan_origin_tls);
 
 	default:
-		UNSUPPORTED("MemorySanitizer used an unrecognized TLS variable: %d", tlsIndex);
+		UNSUPPORTED("MemorySanitizer used an unrecognized TLS variable: %d", static_cast<int>(tlsIndex));
 		return nullptr;
 	}
 }
